chore: move NewAppointmentData, UpdateAppointmentData to types files
